Introduce a 60 req per min limit to Bitfinex API
